+++ Twilio/Tests/Integration/Chat/V2/Service/Channel/MessageTest.php
@@ -49,6 +49,44 @@ class MessageTest extends HolodeckTestCase {
                 "attributes": "{}",
                 "body": "Hello",
                 "index": 0,
+                "type": "text",
+                "media": null,
+                "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+            }
+            '
+        ));
+
+        $actual = $this->twilio->chat->v2->services("ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa")
+                                         ->channels("CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa")
+                                         ->messages("IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa")->fetch();
+
+        $this->assertNotNull($actual);
+    }
+
+    public function testFetchMediaResponse() {
+        $this->holodeck->mock(new Response(
+            200,
+            '
+            {
+                "sid": "IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "service_sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "to": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "channel_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "date_created": "2016-03-24T20:37:57Z",
+                "date_updated": "2016-03-24T20:37:57Z",
+                "was_edited": false,
+                "from": "system",
+                "attributes": "{}",
+                "body": "Hello",
+                "index": 0,
+                "type": "media",
+                "media": {
+                    "sid": "MEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                    "size": 99999999999999,
+                    "content_type": "application/pdf",
+                    "filename": "hello.pdf"
+                },
                 "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '
@@ -100,6 +138,8 @@ class MessageTest extends HolodeckTestCase {
                 "from": "system",
                 "body": "Hello",
                 "index": 0,
+                "type": "text",
+                "media": null,
                 "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '
@@ -129,6 +169,8 @@ class MessageTest extends HolodeckTestCase {
                 "attributes": "{}",
                 "body": "Hello",
                 "index": 0,
+                "type": "text",
+                "media": null,
                 "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '
@@ -185,6 +227,30 @@ class MessageTest extends HolodeckTestCase {
                         "attributes": "{}",
                         "body": "Hello",
                         "index": 0,
+                        "type": "text",
+                        "media": null,
+                        "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                    },
+                    {
+                        "sid": "IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "service_sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "to": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "channel_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "date_created": "2016-03-24T20:37:57Z",
+                        "date_updated": "2016-03-24T20:37:57Z",
+                        "was_edited": false,
+                        "from": "system",
+                        "attributes": "{}",
+                        "body": "Hello",
+                        "index": 0,
+                        "type": "media",
+                        "media": {
+                            "sid": "MEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                            "size": 99999999999999,
+                            "content_type": "application/pdf",
+                            "filename": "hello.pdf"
+                        },
                         "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ]
@@ -287,6 +353,8 @@ class MessageTest extends HolodeckTestCase {
                 "from": "system",
                 "body": "Hello",
                 "index": 0,
+                "type": "text",
+                "media": null,
                 "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '