+++ Twilio/Tests/Integration/IpMessaging/V2/ServiceTest.php
@@ -63,7 +63,11 @@ class ServiceTest extends HolodeckTestCase {
                 "webhook_filters": [
                     "webhook_filters"
                 ],
-                "webhook_method": "webhook_method"
+                "webhook_method": "webhook_method",
+                "media": {
+                    "size_limit_mb": 150,
+                    "compatibility_message": "media compatibility message"
+                }
             }
             '
         ));
@@ -151,7 +155,11 @@ class ServiceTest extends HolodeckTestCase {
                 "webhook_filters": [
                     "webhook_filters"
                 ],
-                "webhook_method": "webhook_method"
+                "webhook_method": "webhook_method",
+                "media": {
+                    "size_limit_mb": 150,
+                    "compatibility_message": "media compatibility message"
+                }
             }
             '
         ));
@@ -243,7 +251,11 @@ class ServiceTest extends HolodeckTestCase {
                         "webhook_filters": [
                             "webhook_filters"
                         ],
-                        "webhook_method": "webhook_method"
+                        "webhook_method": "webhook_method",
+                        "media": {
+                            "size_limit_mb": 150,
+                            "compatibility_message": "media compatibility message"
+                        }
                     }
                 ]
             }
@@ -320,7 +332,11 @@ class ServiceTest extends HolodeckTestCase {
                 "webhook_filters": [
                     "webhook_filters"
                 ],
-                "webhook_method": "webhook_method"
+                "webhook_method": "webhook_method",
+                "media": {
+                    "size_limit_mb": 150,
+                    "compatibility_message": "new media compatibility message"
+                }
             }
             '
         ));