+++ Twilio/Tests/Integration/Preview/Bulkexports/ExportConfigurationTest.php
@@ -39,8 +39,7 @@ class ExportConfigurationTest extends HolodeckTestCase {
                 "enabled": true,
                 "webhook_url": "",
                 "webhook_method": "",
-                "resource_type": "Calls",
-                "email": ""
+                "resource_type": "Calls"
             }
             '
         ));
@@ -71,7 +70,6 @@ class ExportConfigurationTest extends HolodeckTestCase {
             {
                 "url": "https://preview.twilio.com/BulkExports/Exports/Calls/Configuration",
                 "enabled": true,
-                "email": "bogus@twilio.com",
                 "webhook_url": "",
                 "resource_type": "Calls",
                 "webhook_method": ""