+++ Twilio/Tests/Integration/Sync/V1/ServiceTest.php
@@ -45,6 +45,7 @@ class ServiceTest extends HolodeckTestCase {
                     "maps": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Maps"
                 },
                 "sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "unique_name": "unique_name",
                 "url": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "webhook_url": "http://www.example.com",
                 "reachability_webhooks_enabled": false,
@@ -112,6 +113,7 @@ class ServiceTest extends HolodeckTestCase {
                     "maps": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Maps"
                 },
                 "sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "unique_name": "unique_name",
                 "url": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "webhook_url": "http://www.example.com",
                 "reachability_webhooks_enabled": false,
@@ -189,6 +191,7 @@ class ServiceTest extends HolodeckTestCase {
                             "maps": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Maps"
                         },
                         "sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "unique_name": "unique_name",
                         "url": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "webhook_url": "http://www.example.com",
                         "reachability_webhooks_enabled": false,
@@ -233,6 +236,7 @@ class ServiceTest extends HolodeckTestCase {
                     "maps": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Maps"
                 },
                 "sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "unique_name": "unique_name",
                 "url": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "webhook_url": "http://www.example.com",
                 "reachability_webhooks_enabled": false,